\name{nscodData}
\alias{nscodData}
\docType{data}
\title{
nscodData
}
\description{
nscodData
}
\usage{data("nscodData")}
\format{
  The format is:
List of 33
 $ noFleets    
 $ fleetTypes  
 $ sampleTimes 
 $ noYears     
 $ years       
 $ nobs        
 $ idx1        
 $ idx2        
 $ obs         
 $ propMat   
 $ stockMeanWeight          
 $ catchMeanWeight          
 $ natMor                   
 $ landFrac                 
 $ disMeanWeight            
 $ landMeanWeight           
 $ propF                    
 $ propM                    
 $ minAge                   
 $ maxAge                   
 $ maxAgePlusGroup          
 $ keyLogFsta               
 $ corFlag                  
 $ keyLogFpar               
 $ keyQpow                  
 $ keyVarF                  
 $ keyVarLogN               
 $ keyVarObs                
 $ stockRecruitmentModelCode
 $ noScaledYears            
 $ keyScaledYears           
 $ keyParScaledYA           
 $ fbarRange                
}
\details{
...
}
\source{
...
}
\references{
...
}
\examples{
data(nscodData)
## maybe str(nscodData) ; plot(nscodData) ...
}
\keyword{datasets}
