\name{nscodParameters}
\alias{nscodParameters}
\docType{data}
\title{
nscodParameters
}
\description{
nscodParameters
}
\usage{data("nscodParameters")}
\format{
  The format is:
List of 14
 $ logFpar     : num [1:9] -5 -5 -5 -5 -5 -5 -5 -5 -5
 $ logQpow     : num(0) 
 $ logSdLogFsta: num [1:2] -0.7 -0.7
 $ logSdLogN   : num [1:2] -0.35 -0.35
 $ logSdLogObs : num [1:7] -0.35 -0.35 -0.35 -0.35 -0.35 -0.35 -0.35
 $ rec_loga    : num(0) 
 $ rec_logb    : num(0) 
 $ logit_rho   : num 0.5
 $ logScale    : num [1:13] 0 0 0 0 0 0 0 0 0 0 ...
 $ logScaleSSB : num(0) 
 $ logPowSSB   : num(0) 
 $ logSdSSB    : num(0) 
 $ logF        : num [1:6, 1:53] 0 0 0 0 0 0 0 0 0 0 ...
 $ logN        : num [1:6, 1:53] 0 0 0 0 0 0 0 0 0 0 ...
}
\details{
...
}
\source{
...
}
\references{
...
}
\examples{
data(nscodParameters)
## maybe str(nscodParameters) ; plot(nscodParameters) ...
}
\keyword{datasets}
