% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samHL.R
\name{plot.samHL}
\alias{plot.samHL}
\title{Plot function for S3 class \code{"samHL"}}
\usage{
\method{plot}{samHL}(x, ...)
}
\arguments{
\item{x}{An object with S3 class \code{"samHL"}}

\item{\dots}{Additional arguments passed to methods; currently unused.}
}
\description{
Plot the regularization path (regularization parameter versus functional
norm).
}
\details{
The x-axis shows regularization parameters on a log scale. The y-axis shows
the functional norm of each component function.
}
\seealso{
\code{\link{samHL}}
}
