% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samLL.R
\name{plot.samLL}
\alias{plot.samLL}
\title{Plot function for S3 class \code{"samLL"}}
\usage{
\method{plot}{samLL}(x, ...)
}
\arguments{
\item{x}{An object with S3 class \code{"samLL"}}

\item{\dots}{Additional arguments passed to methods; currently unused.}
}
\description{
Plot the regularization path (regularization parameter versus functional
norm).
}
\details{
The x-axis shows regularization parameters on a log scale. The y-axis shows
the functional norm of each component function.
}
\seealso{
\code{\link{samLL}}
}
