% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samQL.R
\name{plot.samQL}
\alias{plot.samQL}
\title{Plot function for S3 class \code{"samQL"}}
\usage{
\method{plot}{samQL}(x, ...)
}
\arguments{
\item{x}{An object with S3 class \code{"samQL"}}

\item{\dots}{Additional arguments passed to methods; currently unused.}
}
\description{
Plot the regularization path (regularization parameter versus functional
norm).
}
\details{
The x-axis shows regularization parameters on a log scale. The y-axis shows
the functional norm of each component function.
}
\seealso{
\code{\link{samQL}}
}
