% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samEL.R
\name{print.samEL}
\alias{print.samEL}
\title{Printing function for S3 class \code{"samEL"}}
\usage{
\method{print}{samEL}(x, ...)
}
\arguments{
\item{x}{An object with S3 class \code{"samEL"}}

\item{\dots}{Additional arguments passed to methods; currently unused.}
}
\description{
Print a summary of an object of class \code{"samEL"}.
}
\details{
The output includes the regularization path length and its degrees of
freedom.
}
\seealso{
\code{\link{samEL}}
}
