% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samHL.R
\name{print.samHL}
\alias{print.samHL}
\title{Printing function for S3 class \code{"samHL"}}
\usage{
\method{print}{samHL}(x, ...)
}
\arguments{
\item{x}{An object with S3 class \code{"samHL"}}

\item{\dots}{Additional arguments passed to methods; currently unused.}
}
\description{
Print a summary of an object of class \code{"samHL"}.
}
\details{
The output includes the regularization path length and its degrees of
freedom.
}
\seealso{
\code{\link{samHL}}
}
