% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samLL.R
\name{print.samLL}
\alias{print.samLL}
\title{Printing function for S3 class \code{"samLL"}}
\usage{
\method{print}{samLL}(x, ...)
}
\arguments{
\item{x}{An object with S3 class \code{"samLL"}}

\item{\dots}{Additional arguments passed to methods; currently unused.}
}
\description{
Print a summary of an object of class \code{"samLL"}.
}
\details{
The output includes the regularization path length and its degrees of
freedom.
}
\seealso{
\code{\link{samLL}}
}
