% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samQL.R
\name{print.samQL}
\alias{print.samQL}
\title{Printing function for S3 class \code{"samQL"}}
\usage{
\method{print}{samQL}(x, ...)
}
\arguments{
\item{x}{An object with S3 class \code{"samQL"}}

\item{\dots}{Additional arguments passed to methods; currently unused.}
}
\description{
Print a summary of an object of class \code{"samQL"}.
}
\details{
The output includes the regularization path length and its degrees of
freedom.
}
\seealso{
\code{\link{samQL}}
}
