% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run.R
\name{sam.fit}
\alias{sam.fit}
\title{Fit SAM model}
\usage{
sam.fit(data, parameters)
}
\arguments{
\item{data}{...}

\item{parameters}{...}
}
\description{
Fit SAM model
}
\details{
...
}
\examples{
data(nscodData)
data(nscodParameters)
fit <- sam.fit(nscodData, nscodParameters)
}

