% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reading.R
\name{setup.sam.data}
\alias{setup.sam.data}
\title{Combine the data sources to SAM readable object}
\usage{
setup.sam.data(fleets = NULL, surveys = NULL, residual.fleet = NULL,
  prop.mature = NULL, stock.mean.weight = NULL, catch.mean.weight = NULL,
  dis.mean.weight = NULL, land.mean.weight = NULL,
  natural.mortality = NULL, prop.f = NULL, prop.m = NULL,
  land.frac = NULL)
}
\arguments{
\item{too}{many to list now}
}
\description{
Combine the data sources to SAM readable object
}
\details{
...
}

