% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samHL.R
\name{plot.samHL}
\alias{plot.samHL}
\title{Plot function for S3 class \code{"samHL"}}
\usage{
\method{plot}{samHL}(x, ...)
}
\arguments{
\item{x}{An object with S3 class \code{"samHL"}}

\item{\dots}{System reserved (No specific usage)}
}
\description{
This function plots the regularization path (regularization parameter versus functional norm)
}
\details{
The horizontal axis is for the regularization parameters in log scale. The vertical axis is for the functional norm of each component.
}
\seealso{
\code{\link{samHL}}
}
