% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samEL.R
\name{predict.samEL}
\alias{predict.samEL}
\title{Prediction function for S3 class \code{"samEL"}}
\usage{
\method{predict}{samEL}(object, newdata, ...)
}
\arguments{
\item{object}{An object with S3 class \code{"samEL"}.}

\item{newdata}{The testing dataset represented in a \code{n} by \code{d} matrix, where \code{n} is testing sample size and \code{d} is dimension.}

\item{\dots}{System reserved (No specific usage)}
}
\value{
\item{expectations}{
    Estimated expected counts also represented in a \code{n} by the length of \code{lambda} matrix, where \code{n} is testing sample size.
  }
}
\description{
Predict the labels for testing data.
}
\details{
The testing dataset is rescale to the samELe range, and expanded by the samELe spline basis functions as the training data.
}
\seealso{
\code{\link{samEL}}
}
