% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samQL.R
\name{predict.samQL}
\alias{predict.samQL}
\title{Prediction function for S3 class \code{"samQL"}}
\usage{
\method{predict}{samQL}(object, newdata, ...)
}
\arguments{
\item{object}{An object with S3 class \code{"samQL"}.}

\item{newdata}{The testing dataset represented in a \code{n} by \code{d} matrix, where \code{n} is testing sample size and \code{d} is dimension.}

\item{\dots}{System reserved (No specific usage)}
}
\value{
\item{values}{
  Predicted values also represented in a \code{n} by the length of \code{lambda} matrix, where \code{n} is testing sample size.
}
}
\description{
Predict the labels for testing data.
}
\details{
The testing dataset is rescale to the samQLe range, and expanded by the samQLe spline basis functions as the training data.
}
\seealso{
\code{\link{samQL}}
}
