% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samEL.R
\name{print.samEL}
\alias{print.samEL}
\title{Printing function for S3 class \code{"samEL"}}
\usage{
\method{print}{samEL}(x, ...)
}
\arguments{
\item{x}{An object with S3 class \code{"samEL"}}

\item{\dots}{System reserved (No specific usage)}
}
\description{
Summarize the information of the object with S3 class \code{samEL}.
}
\details{
The output includes length and d.f. of the regularization path.
}
\seealso{
\code{\link{samEL}}
}
