% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samHL.R
\name{print.samHL}
\alias{print.samHL}
\title{Printing function for S3 class \code{"samHL"}}
\usage{
\method{print}{samHL}(x, ...)
}
\arguments{
\item{x}{An object with S3 class \code{"samHL"}}

\item{\dots}{System reserved (No specific usage)}
}
\description{
Summarize the information of the object with S3 class \code{samHL}.
}
\details{
The output includes length and d.f. of the regularization path.
}
\seealso{
\code{\link{samHL}}
}
