% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samLL.R
\name{print.samLL}
\alias{print.samLL}
\title{Printing function for S3 class \code{"samLL"}}
\usage{
\method{print}{samLL}(x, ...)
}
\arguments{
\item{x}{An object with S3 class \code{"samLL"}}

\item{\dots}{System reserved (No specific usage)}
}
\description{
Summarize the information of the object with S3 class \code{samLL}.
}
\details{
The output includes length and d.f. of the regularization path.
}
\seealso{
\code{\link{samLL}}
}
