% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samQL.R
\name{print.samQL}
\alias{print.samQL}
\title{Printing function for S3 class \code{"samQL"}}
\usage{
\method{print}{samQL}(x, ...)
}
\arguments{
\item{x}{An object with S3 class \code{"samQL"}}

\item{\dots}{System reserved (No specific usage)}
}
\description{
Summarize the information of the object with S3 class \code{samQL}.
}
\details{
The output includes length and d.f. of the regularization path.
}
\seealso{
\code{\link{samQL}}
}
